Frame Rate Update
Now you can change the frameRate() using the u/U (increase) and i/I (decrease) keys on any page. So it can make things easier to showcase and see what speed I want this to be later on.

I also made notes on how to change the p5.js canvas to full screen.